+++ Examples/Nodify.Calculator/Operations/UnaryOperation.cs
@@ -4,11 +4,11 @@ namespace Nodify.Calculator
 {
     public class UnaryOperation : IOperation
     {
-        private readonly Func<decimal, decimal> _func;
+        private readonly Func<double, double> _func;
 
-        public UnaryOperation(Func<decimal, decimal> func) => _func = func;
+        public UnaryOperation(Func<double, double> func) => _func = func;
 
-        public decimal Execute(params decimal[] operands)
+        public double Execute(params double[] operands)
             => _func.Invoke(operands[0]);
     }
 }
